refactor: extract CrossCheckComments component from cross-check-submit page
