+++ client/src/components/index.tsx
@@ -1,4 +1,5 @@
 export * from './AdminSider';
+export * from './CrossCheckComments';
 export * from './Footer';
 export * from './GithubAvatar';
 export * from './GithubUserLink';
