+++ client/src/pages/course/student/cross-check-submit.tsx
@@ -1,6 +1,5 @@
-import { UserOutlined } from '@ant-design/icons';
-import { Button, Col, Comment, Alert, Divider, Form, Input, message, Row, Typography } from 'antd';
-import { PageLayout, GithubUserLink } from 'components';
+import { Button, Col, Alert, Form, Input, message, Row, Typography } from 'antd';
+import { PageLayout, CrossCheckComments } from 'components';
 import withCourseData from 'components/withCourseData';
 import withSession from 'components/withSession';
 import { useMemo, useState } from 'react';
@@ -99,43 +98,6 @@ function Page(props: CoursePageProps) {
 }
 
 export default withCourseData(withSession(Page, 'student'));
-
-function CrossCheckComments({
-  comments,
-}: {
-  comments: { comment: string; score: number; author: { name: string; githubId: string } | null }[];
-}) {
-  if (!comments || comments.length === 0) {
-    return null;
-  }
-  const style = { margin: 16, fontStyle: 'italic' };
-  return (
-    <Col>
-      {comments.map(({ comment, author, score }, i) => (
-        <Row key={i}>
-          <Divider />
-          <Comment
-            style={style}
-            author={author ? <GithubUserLink value={author.githubId} /> : `Student ${i + 1}`}
-            avatar={<UserOutlined />}
-            content={
-              <>
-                {score ? (
-                  <p>
-                    <b>Score: {score}</b>
-                  </p>
-                ) : null}
-                {comment.split('\n').map((text, k) => (
-                  <p key={k}>{text}</p>
-                ))}
-              </>
-            }
-          />
-        </Row>
-      ))}
-    </Col>
-  );
-}
 
 function renderDeadlineInfo(isSubmitDisabled: boolean) {
   return (
